feat: Move agnostic upgrade to upgrade_helpers
