+++ src/upgrades/machine_upgrade.py
@@ -12,33 +12,34 @@ import time
 import typing
 
 import ops
+from charms.mongodb.v0.upgrade_helpers import AbstractUpgrade, UnitState
 
 from config import Config
-from upgrades import mongodb_upgrade, upgrade
+from upgrades import mongodb_upgrade
 
 logger = logging.getLogger(__name__)
 
 _SNAP_REVISION = str(Config.SNAP_PACKAGES[0][2])
 
 
-class Upgrade(upgrade.Upgrade):
+class Upgrade(AbstractUpgrade):
     """In-place upgrades on machines."""
 
     @property
-    def unit_state(self) -> typing.Optional[upgrade.UnitState]:
+    def unit_state(self) -> typing.Optional[UnitState]:
         """Returns the unit state."""
         if (
             self._unit_workload_container_version is not None
             and self._unit_workload_container_version != self._app_workload_container_version
         ):
             logger.debug("Unit upgrade state: outdated")
-            return upgrade.UnitState.OUTDATED
+            return UnitState.OUTDATED
         return super().unit_state
 
     @unit_state.setter
-    def unit_state(self, value: upgrade.UnitState) -> None:
+    def unit_state(self, value: UnitState) -> None:
         # Super call
-        upgrade.Upgrade.unit_state.fset(self, value)
+        AbstractUpgrade.unit_state.fset(self, value)
 
     def _get_unit_healthy_status(self) -> ops.StatusBase:
         if self._unit_workload_container_version == self._app_workload_container_version:
@@ -151,11 +152,11 @@ class Upgrade(upgrade.Upgrade):
                 return True
             state = self._peer_relation.data[unit].get("state")
             if state:
-                state = upgrade.UnitState(state)
+                state = UnitState(state)
             if (
                 self._unit_workload_container_versions.get(unit.name)
                 != self._app_workload_container_version
-                or state is not upgrade.UnitState.HEALTHY
+                or state is not UnitState.HEALTHY
             ):
                 # Waiting for higher number units to upgrade
                 return False
@@ -179,7 +180,7 @@ class Upgrade(upgrade.Upgrade):
             return
 
         logger.debug(f"Upgrading {self.authorized=}")
-        self.unit_state = upgrade.UnitState.UPGRADING
+        self.unit_state = UnitState.UPGRADING
         charm.install_snap_packages(packages=Config.SNAP_PACKAGES)
         self._unit_databag["snap_revision"] = _SNAP_REVISION
         self._unit_workload_version = self._current_versions["workload"]
